Correct Namespaces
Some tests were in the wrong location with the wrong namespaces.
These have been corrected.
